==> RCT-PPMM/07_simulation_plots_PPMM_MLE_withTrialBounds.Rmd <==
---
title: "Simulation - Plots - MLEs"
author: "Rebecca Andridge"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE, out.width='100%')
library(tidyverse)
theme_set(theme_bw())

# load results -----
load("./simdata/simParms.RData")
load("./simresults/simSummary_PPMM_MLE.RData")
load("./simresults/simSummary_PPMM_MLE_cover.RData")
load("./simresults/simSummary_truth.RData")
load("./simresults/simSummary_trial.RData")

parms <- parms %>%
  mutate(simset = factor(simset),
         S2_f = factor(S2, levels = c(13,4,1), labels = c("weak proxy", "moderate proxy", "strong proxy")),
         effectmod = factor(effectmod,
                            levels = c("EM{U}", "EM{Z,U}", "EM{Z,W,U}")),
         selection = factor(selection, 
                            levels = c("SEL{U}", "SEL{Z,U}", "SEL{Z,W,U}", "SEL{Z,-U}", "SEL{Z,W,-U}")),
         CORR_U_W = factor(CORR_U_W, levels=c(0.3,0)))

# Trial estimates
trial_est <- sumstats_trial %>%
  filter(stat == "thetamean") %>%
  select(simset, estimator, est)
trial_empse <- sumstats_trial %>%
  filter(stat == "empse") %>%
  select(simset, est) %>%
  rename(empse = est)
trial_est <- trial_est %>%
  full_join(trial_empse) %>%
  mutate(lb = est - qt(0.975, df = 198)*empse,
         ub = est + qt(0.975, df = 198)*empse) %>%
  full_join(parms) %>% full_join(truth)
rm(sumstats_trial, trial_empse)

# PPMM estimates -----
est <- sumstats_mle %>%
  filter(stat == "thetamean") %>%
  inner_join(truth) %>%
  full_join(parms)
rm(sumstats_mle)
```

# Estimates of Treatment Effect

```{r}
est_wide1 <- est %>%
  filter(estimator %in% c("trt_phi0_pla_phi0",
                          "trt_phi0.5_pla_phi0",
                          "trt_phi1_pla_phi0")) %>%
  select(stat, est, estimator, simset, S2, effectmod, selection, sel_neg_u, CORR_U_W, truth) %>%
  pivot_wider(names_from = estimator, values_from = est) %>%
  rename(lb  = trt_phi0_pla_phi0,
         mid = trt_phi0.5_pla_phi0,
         ub  = trt_phi1_pla_phi0) %>%
  add_column(type = "pla_phi_0")
est_wide2 <- est %>%
  filter(estimator %in% c("trt_phi0_pla_phi0",
                          "trt_phi0_pla_phi0.5",
                          "trt_phi0_pla_phi1")) %>%
  select(stat, est, estimator, simset, S2, effectmod, selection, sel_neg_u, CORR_U_W, truth) %>%
  pivot_wider(names_from = estimator, values_from = est) %>%
  rename(lb  = trt_phi0_pla_phi0,
         mid = trt_phi0_pla_phi0.5,
         ub  = trt_phi0_pla_phi1) %>%
  add_column(type = "trt_phi_0")
trial <- trial_est %>%
  select(simset, est, lb, ub, S2, effectmod, selection, sel_neg_u, CORR_U_W, truth) %>%
  rename(mid = est) %>%
  add_column(type = "trial") %>%
  mutate(cover = ub > truth & lb < truth)
est_wide <- bind_rows(est_wide1, est_wide2, trial) %>%
  mutate(type = factor(type,
                       levels = c("trt_phi_0", "pla_phi_0", "trial"),
                       labels = c("RCT-PPMM: Selection depends on Control Outcome",
                                  "RCT-PPMM: Selection depends on Treatment Outcome",
                                  "Trial"))) %>%
  mutate(S2_f = factor(S2, levels=c(1,4,13), labels=c("strong proxy", "moderate proxy", "weak proxy")))
rm(est_wide1, est_wide2, trial, om)
```

```{r}
# Plot
pd <- position_dodge(0.5)

g <- est_wide %>%
  ggplot(aes(y=CORR_U_W, x=mid, col=type, shape=type)) + 
  geom_point(size=1, position=pd) + 
  geom_errorbar(aes(xmin = lb, xmax = ub), width=0.3, position=pd) +
  facet_grid(effectmod ~ selection, scales = "free_x") + 
  geom_point(aes(y=CORR_U_W, x=truth, col=type, shape=type),col="black", size=5, shape="|") +
  labs(y="Cor(U,W)", x="Estimated Treatment Effect", color="", shape="") +
  theme(legend.position="top") + 
  scale_color_manual(values = c("blue", "forestgreen", "red"),
                     guide = guide_legend(reverse=TRUE)) +
  scale_shape_manual(values = c(15,16,17),
                     guide = guide_legend(reverse=TRUE))
```

### Strong Proxy

Black vertical line is truth

```{r}
g %+% 
  filter(est_wide, S2 == 1) 

ggsave("./plots/sim_trtest_trialBounds_proxySTRONG.pdf", width=8, height=5)
ggsave("./plots/sim_trtest_trialBounds_proxySTRONG.png", width=8, height=5)
```

### Moderate Proxy

Black vertical line is truth

```{r}
g %+% 
  filter(est_wide, S2 == 4) 

ggsave("./plots/sim_trtest_trialBounds_proxyMODERATE.pdf", width=8, height=5)
ggsave("./plots/sim_trtest_trialBounds_proxyMODERATE.png", width=8, height=5)
```

### Weak Proxy

Black vertical line is truth

```{r}
g %+% 
  filter(est_wide, S2 == 13) 

ggsave("./plots/sim_trtest_trialBounds_proxyWEAK.pdf", width=8, height=5)
ggsave("./plots/sim_trtest_trialBounds_proxyWEAK.png", width=8, height=5)
```
